Adapt structure according to light-poseidon and Make workspace functional upon checkout
