+++ circuit-lib/circuit-lib.rs/decode-babyjubjub/.gitignore
@@ -1,0 +1,4 @@
+target
+Cargo.lock
+src/*.bincode
+!decode_lookup_table16.bincode
