--- circuit-lib/circuit-lib.rs/src/decode-babyjubjub/decode/.gitignore
@@ -1,3 +1,0 @@
-target
-Cargo.lock
-src/*.bincode
