+++ circuit-lib/circuit-lib.rs/xtask/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 bincode = "1.3.3"
-decode = { path = "../babyjubjub_ecdlp"}
+decode = { path = "../decode-babyjubjub"}
 indicatif = "0.17.7"
 clap = { version = "4.4.5", features = ["derive"] }
 anyhow = "1.0.75"
